Make scalismo work with Scala 3 (#374)
This commit adds support for Scala 3 and allows cross-compilation with Scala 2.13. Support for Scala 2.12 was removed.
Besides Scala 3 support, the commit also updates all major dependencies, most notably breeze.

+++ src/test/scala/scalismo/utils/MemoizationTests.scala
@@ -34,7 +34,7 @@ class MemoizationTests extends ScalismoTestSuite {
       def time[A](a: () => A): Double = {
         val now = System.nanoTime
         a()
-        System.nanoTime() - now
+        (System.nanoTime() - now).toDouble
       }
 
       def slowTestFun(x: Double) = {
